test: init migration
